add tested comments, move parseServerSide*Params to pages and export all utils
