+++ packages/next-typesafe-url/src/index.ts
@@ -33,6 +33,7 @@ export type {
   ServerParseParamsResult,
 };
 
+// * TESTED
 export function $path<T extends AllRoutes>({
   route,
   searchParams,
